std/stream: Refactor into multiple files.
